% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compareTests.R
\name{compareTests}
\alias{compareTests}
\title{compareTests}
\usage{
compareTests(x, softwares = names(x), versions = lapply(x, names),
  ref.software = NULL, ref.versions = NULL, mapped.tests = NULL,
  show.all.tests = is.null(mapped.tests))
}
\arguments{
\item{x}{param description}

\item{softwares}{param description}

\item{versions}{param description}

\item{ref.software}{param description}

\item{ref.versions}{param description}

\item{mapped.tests}{param description}

\item{show.all.tests}{param description}
}
\value{
some comparison results
}
\description{
Compares Tests.
}
\details{
Current version take library of test results from compileTestResults().  could write generic function that take result from above as argument or that can accept all the data and run the above too.
}

