% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compileTestResults.R
\name{compileTestResults}
\alias{compileTestResults}
\title{compileTestResults}
\usage{
compileTestResults(softwareList, outputsDirTop)
}
\arguments{
\item{softwareList}{list of softwares}

\item{outputsDirTop}{top of output dirs}
}
\value{
a list of lists of lists [[library]][[software]][[version]][[test]]
}
\description{
Compiles test results.
}
\details{
creates a list of lists (of lists) [[library]][[software]][[version]][[test]]
}

