% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertLinkToRelative.R
\name{convertLinkToRelative}
\alias{convertLinkToRelative}
\title{convertLinkToRelative}
\usage{
convertLinkToRelative(x, start.dir = getwd())
}
\arguments{
\item{x}{a path to a file}

\item{start.dir}{from which directory should the relative link be made [default working dir]}
}
\value{
a character string relative path
}
\description{
convertLinkToRelative
}
\details{
converts a file path to a relative link within the system. If the file is on a different system, it is unchanged.
}

