% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/createTestLinks.R
\name{createTestLinks}
\alias{createTestLinks}
\title{createTestLinks}
\usage{
createTestLinks(test, software, testFolder, format = "markdown",
  check.test.file = TRUE, relative.link = TRUE, start.dir = getwd())
}
\arguments{
\item{test}{a data frame of the type returned by compareTests()}

\item{software}{which software}

\item{testFolder}{top level testfolder}

\item{format}{c("markdown", "html")}

\item{check.test.file}{check whether the created link points to a real file}

\item{relative.link}{make the file path relative (uses convertLinkToRelative())}

\item{start.dir}{to which directory should the link be relative}
}
\description{
createTestLinks
}
\details{
creates links to test file locations to be used in html/markdown reporting.
}

