% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findTestOutputs.R
\name{findTestOutputs}
\alias{findTestOutputs}
\title{try to find expected output from test files.}
\usage{
findTestOutputs(testDir)
}
\arguments{
\item{testDir}{the directory containing tests and outputs}
}
\value{
a vector of named test results.
}
\description{
try to find expected output from test files. (long)
}
\details{
in bash, should use the TEST_OUT_NAME variable.
}

