% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Kanute.R
\docType{package}
\name{Kanute-package}
\alias{Kanute-package}
\title{Kanute: unit-style behaviour testing for softwares and software versions.}
\description{
\tabular{ll}{
Package: \tab Kanute\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-01-31\cr
Author: \tab Dave T. Gerrard \cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\details{
Test functionality of software as per unit tests.
Compare functionality between versions of same software.
Compare functionality between different software (given mapping of functionality).
More information about each function can be found in
its help documentation.

Some text

The main functions are:
\code{\link{compileTestResults}}, \code{\link{compareTests}},
\code{\link{scanTestDir}}, \code{\link{findTestOutputs}},

...
}

