% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.TestLibrary.R
\name{plot.testLibrary}
\alias{plot.testLibrary}
\title{plot.testLibrary}
\usage{
\method{plot}{testLibrary}(x = test.library, softwares = names(x),
  versions = lapply(x, names), ref.software = NULL, ref.versions = NULL,
  mapped.tests = NULL, show.all.tests = is.null(mapped.tests), ...)
}
\arguments{
\item{x}{param description}

\item{softwares}{param description}

\item{versions}{param description}

\item{ref.software}{param description}

\item{ref.versions}{param description}

\item{mapped.tests}{param description}

\item{show.all.tests}{param description}
}
\description{
Plot all tests in test.library
}
\details{
Not yet implemented
}

