% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.versionTest.R
\name{plot.versionTest}
\alias{plot.versionTest}
\title{plot.versionTest}
\usage{
\method{plot}{versionTest}(x, versions = setdiff(names(x), "test"))
}
\arguments{
\item{x}{a data frame of the type returned by compareTests()}

\item{versions}{Which software versions to include (default is all).}
}
\description{
plot.versionTest
}
\details{
Plots the results of compareTests() for an individual piece of software.
}

